xlat: Allow regions to specify their granularity
Allow individual regions to request a finer granularity than the one
normally assigned by the translation tables library. The library tries
to save as much memory as possible, which may not be what the user wants
in some cases, like in cases when a region needs to be split into
subregions with different attributes.

Change-Id: I04de9302e0ee3d326b8877043a9f638766b81b7b
Signed-off-by: Antonio Nino Diaz <antonio.ninodiaz@arm.com>