SPM: Improve support for managing partition state
State information is maintained for a secure partition only on the primary CPU
while it is being initialised. This patch implements support to track the state
of a partition on a per-cpu basis and also extends the number of states that are
tracked. This support will be used in subsequent patches to initialise secure
partitions on secondary CPUs.

Change-Id: If9862ee3e5e263ec03728bed05a27fe6bc244362
Signed-off-by: Achin Gupta <achin.gupta@arm.com>