SPM: Move secure partition context initialisation
The S-EL0 specific context for a secure partition was being initialised
immediately before jumping into it. This approach works if the partition is to
be initialised but not if the partition has already been initialised but still
needs to be entered. This patch moves the corresponding function from common
code to initialisation specific code.

Change-Id: Ice285fd134dc154bca9db1ed3b3ee7dc1dd28a8c
Signed-off-by: Achin Gupta <achin.gupta@arm.com>