SPM: Initialise secure partitions on secondary cpus
This patch leverages the SPD power management operations structure to register
notifiers for PSCI CPU_ON, CPU_OFF and CPU_SUSPEND operations. The notifier for
the ON finish operation enables the SPM to initialise the secure partition on
secondary CPUs. The notifiers for all other operations are placeholders for
future expansion.

Some private APIs have been made public to enable re-use in the notifier
functions.

An interface has been added to initialise the context for a secondary CPU during
a warm boot that enables it to jump into the Secure partition to perform any
initialisations.

A data structure has been defined that is used to pass boot information to the
secondary CPU.

Change-Id: I0b4407e551eefe18f2bad3947e0262fee0568f01
Signed-off-by: Achin Gupta <achin.gupta@arm.com>